moves the drivers folder to the infrastructure layer to follow the hexagonal architecture approach
