+++ src/main/java/com/fixed_asset/appasset/infrastructure/http_rest/PersonController.java
@@ -1,20 +1,13 @@
-package com.fixed_asset.appasset.controller;
+package com.fixed_asset.appasset.infrastructure.http_rest;
 
-import com.fixed_asset.appasset.domain.Area;
-import com.fixed_asset.appasset.domain.Asset;
 import com.fixed_asset.appasset.domain.Person;
-import com.fixed_asset.appasset.domain.services.AreaServices;
-import com.fixed_asset.appasset.domain.services.AssetServices;
 import com.fixed_asset.appasset.domain.services.PersonServices;
 import com.fixed_asset.appasset.exceptions.ApiException;
-import com.fixed_asset.appasset.exceptions.BadRequestException;
-import com.fixed_asset.appasset.exceptions.ValidationError;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.time.LocalDateTime;
 import java.util.List;
